Add files via upload
Donor+  is an android application which is implemented using Java and Mysql. 
This application is created to help people who need blood. A user must first create an account. Upon logging in, a user can read posts of people who need blood. He can call a user or share his post. A user can also add a post or search for donors belonging to a specific blood group and city.